refactor: split shoppingCart slice into shoppingCart and shoppingCartUi slices and update all components using shoppingCart
